Splits topup covariates in two separate files.
